+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/EventMapping.kt
@@ -254,6 +254,7 @@ private fun MemberRemovedEventDto.toDomain(): MemberRemovedEvent {
         cid = cid,
         channelType = channel_type,
         channelId = channel_id,
+        member = member.toDomain(),
     )
 }