+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/utils/DevTokenTest.kt
@@ -4,6 +4,7 @@ import com.nhaarman.mockitokotlin2.mock
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.clientstate.SocketStateService
 import io.getstream.chat.android.client.clientstate.UserStateService
+import io.getstream.chat.android.client.experimental.errorhandler.factory.NoOpErrorHandlerFactory
 import io.getstream.chat.android.client.helpers.QueryChannelsPostponeHelper
 import io.getstream.chat.android.client.token.FakeTokenManager
 import io.getstream.chat.android.client.utils.retry.NoRetryPolicy
@@ -34,6 +35,7 @@ internal class DevTokenTest(private val userId: String, private val expectedToke
         userStateService = userStateService,
         scope = testCoroutines.scope,
         retryPolicy = NoRetryPolicy(),
+        errorHandlerFactory = NoOpErrorHandlerFactory(),
     )
 
     @Test