+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/common/composer/MessageComposerController.kt
@@ -177,6 +177,11 @@ public class MessageComposerController(
         get() = messageMode.value is MessageMode.MessageThread
 
     /**
+     * Represents the selected mentions based on the message suggestion list.
+     */
+    private val selectedMentions: MutableSet<User> = mutableSetOf()
+
+    /**
      * Sets up the data loading operations such as observing the maximum allowed message length.
      */
     init {
@@ -356,7 +361,7 @@ public class MessageComposerController(
      * Clears all the data from the input - both the current [input] value and the
      * [selectedAttachments].
      */
-    private fun clearData() {
+    public fun clearData() {
         input.value = ""
         selectedAttachments.value = emptyList()
         validationErrors.value = emptyList()
@@ -411,11 +416,13 @@ public class MessageComposerController(
 
         val actionMessage = activeAction?.message ?: Message()
         val replyMessageId = (activeAction as? Reply)?.message?.id
+        val mentions = filterMentions(selectedMentions, message)
 
         return if (isInEditMode) {
             actionMessage.copy(
                 text = message,
-                attachments = attachments.toMutableList()
+                attachments = attachments.toMutableList(),
+                mentionedUsersIds = mentions
             )
         } else {
             Message(
@@ -423,9 +430,30 @@ public class MessageComposerController(
                 text = message,
                 parentId = parentMessageId,
                 replyMessageId = replyMessageId,
-                attachments = attachments.toMutableList()
+                attachments = attachments.toMutableList(),
+                mentionedUsersIds = mentions
             )
         }
+    }
+
+    /**
+     * Filters the current input and the mentions the user selected from the suggestion list. Removes any mentions which
+     * are selected but no longer present in the input.
+     *
+     * @param selectedMentions The set of selected users from the suggestion list.
+     * @param message The current message input.
+     *
+     * @return [MutableList] of user IDs of mentioned users.
+     */
+    private fun filterMentions(selectedMentions: Set<User>, message: String): MutableList<String> {
+        val text = message.lowercase()
+
+        val remainingMentions = selectedMentions.filter {
+            text.contains("@${it.name.lowercase()}")
+        }.map { it.id }
+
+        this.selectedMentions.clear()
+        return remainingMentions.toMutableList()
     }
 
     /**
@@ -510,6 +538,7 @@ public class MessageComposerController(
         val augmentedMessageText = "${messageText.substringBeforeLast("@")}@${user.name} "
 
         setMessageInput(augmentedMessageText)
+        selectedMentions += user
     }
 
     /**