+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/viewholder/decorator/internal/BaseDecorator.kt
@@ -7,6 +7,7 @@ import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.Dat
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.GiphyAttachmentViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.GiphyViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.ImageAttachmentViewHolder
+import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.LinkAttachmentsViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.MessageDeletedViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.MessagePlainTextViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.TextAndAttachmentsViewHolder
@@ -24,6 +25,7 @@ internal abstract class BaseDecorator : Decorator {
             is MessageDeletedViewHolder -> decorateDeletedMessage(viewHolder, data)
             is MessagePlainTextViewHolder -> decoratePlainTextMessage(viewHolder, data)
             is TextAndAttachmentsViewHolder -> decorateTextAndAttachmentsMessage(viewHolder, data)
+            is LinkAttachmentsViewHolder -> decorateLinkAttachmentsMessage(viewHolder, data)
             is GiphyViewHolder -> decorateGiphyMessage(viewHolder, data)
             is GiphyAttachmentViewHolder -> decorateGiphyAttachmentMessage(viewHolder, data)
             is ImageAttachmentViewHolder -> decorateImageAttachmentMessage(viewHolder, data)
@@ -61,6 +63,17 @@ internal abstract class BaseDecorator : Decorator {
 
     protected abstract fun decoratePlainTextMessage(
         viewHolder: MessagePlainTextViewHolder,
+        data: MessageListItem.MessageItem,
+    )
+
+    /**
+     * Applies various decorations to [LinkAttachmentsViewHolder].
+     *
+     * @param viewHolder The holder to be decorated.
+     * @param data The data used to define various decorations.
+     */
+    abstract fun decorateLinkAttachmentsMessage(
+        viewHolder: LinkAttachmentsViewHolder,
         data: MessageListItem.MessageItem,
     )